lint & fix imports from rename
